Heuristic new

+++ src/engine/utils.rs
@@ -46,10 +46,10 @@ pub fn random_bits(n: u8) -> u128 {
 
 
 #[inline(always)]
-pub fn natural_log(x: u32) -> f32 {
+pub fn natural_log(x: f32) -> f32 {
     // TODO optimize. search "fast natural log"
-    if x >= N_NATURAL_LOGS as u32 {
-        return (x as f32).log(2.71828182845);
+    if x >= N_NATURAL_LOGS as f32 {
+        return x.log(2.71828182845);
     }
     unsafe {
         return NATURAL_LOG_TABLE[x as usize];
